Debugged
Debugged the work from yesterday, one Bug remaining but it doesn't break anything, and no real apparent reason for the error.
LayerID Line 153 apparently isn't int?
